+++ spec/models/conversion/task_list_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Conversion::TaskList do
         :conversion_grant,
         :sponsored_support_grant,
         :academy_details,
-        :funding_agreement_contact,
+        :main_contact,
         :land_questionnaire,
         :land_registry,
         :supplemental_funding_agreement,
@@ -55,7 +55,7 @@ RSpec.describe Conversion::TaskList do
               Conversion::Task::ConversionGrantTaskForm,
               Conversion::Task::SponsoredSupportGrantTaskForm,
               Conversion::Task::AcademyDetailsTaskForm,
-              Conversion::Task::FundingAgreementContactTaskForm
+              Conversion::Task::MainContactTaskForm
             ]
           },
           {