+++ spec/spec_helper.rb
@@ -19,6 +19,7 @@ SimpleCov.minimum_coverage 100
 unless ENV.fetch("NO_COVERAGE", false) == "true"
   SimpleCov.start "rails" do
     add_filter "lib/generators/"
+    add_filter "app/forms/conversion/task/funding_agreement_contact_task_form.rb"
   end
 end